mcp_server/requirements.txt
# Hemingwix Notion MCP Server Requirements

# MCP (Model Context Protocol)
mcp>=1.0.0

# Notion API
notion-client>=2.0.0

# Core dependencies
asyncio-mqtt>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Enhanced logging and debugging
structlog>=23.0.0
rich>=13.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0